### 2024-11-12 #7457 @editorialbot
[PRE REVIEW]: Taxonomy Resolver: A Python package for building and filtering taxonomy trees
**Submitting author:** @biomadeira (<a href="http://orcid.org/0000-0001-8728-9449">Fabio Madeira</a>)
**Repository:** https://github.com/ebi-jdispatcher/taxonomy-resolver
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
