### 2024-12-17 #7457 @Kevin-Mattheus-Moerman
@biomadeira yes, I am happy to report this has passed the initial scope review. I will now proceed to try to find a handling editor.
