### 2024-12-18 #7457 @csoneson
👋🏻 @biomadeira I will handle your submission, and start by finding two suitable reviewers. If you have suggestions, feel free to mention them here (see for example the list in the first post above).
In the meantime, could you take a look at the DOI issues flagged above? In particular, to fix the invalid DOI (it should not contain the 'https://doi.org/' prefix) and to double-check if there is a DOI for the other mentioned references. Thanks!
