### 2024-12-18 #7457 @biomadeira
Hi @csoneson!

ebi-jdispatcher/taxonomy-resolver#11 should have fixed the DOI issue
