### 2024-12-18 #7457 @SimonGreenhill
Looks interesting, but I'm swamped with end of year things at the moment and on leave for most of January (parenting in school holidays).
