### 2024-12-19 #7457 @ryneches
Oh hey, I've written, like, five crappy, _ad hoc_ versions of this tool over the years. Glad someone took a stab at doing it properly.

Sure, I'll review!
