### 2024-12-19 #7457 @ggonnella
I have some holiday very soon and when I am back, I have some urgent
obligations, so I can't accept right now, unfortunately...

On Thu, Dec 19, 2024 at 1:21 PM Russell Neches ***@***.***>
wrote:

> Oh hey, I've written, like, five crappy, *ad hoc* versions of this tool
> over the years. Glad someone took a stab at doing it properly.
>
> Sure, I'll review!
>
