### 2024-12-19 #7604 @editorialbot
[REVIEW]: Taxonomy Resolver: A Python package for building and filtering taxonomy trees
**Submitting author:** @biomadeira (<a href="http://orcid.org/0000-0001-8728-9449">Fabio Madeira</a>)
**Repository:** https://github.com/ebi-jdispatcher/taxonomy-resolver
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.1
