### 2024-12-19 363af39
paper: couple of small fixes
