### 2024-12-20 5b58dec
fix typo in CLI docs
