### 2024-12-20 ebi-jdispatcher/taxonomy-resolver#13 @ryneches
Bring test data up to date
The test data is based on a snapshot of the NCBI database that's four years out of date. It still works as long as you use the `taxdmp.zip` file provided in the `testdata` directory. However, if you download a fresh snapshot of the database and build a new tree (as the CLI instructions suggest you do), you run into some problems :

```
$ taxonomy-resolver build -in tree.pickle -inf pickle -out tree_filtered.pickle -outf pickle -taxidsf testdata/taxids_filter.txt
20/12/2024 13:35:02 - [INFO] Validated inputs and outputs.
20/12/2024 13:35:02 - [INFO] Loaded NCBI Taxonomy from 'tree.pickle' in 'pickle' format.
Some of the provided TaxIDs are not valid or not found in the built Tree.
```

Unless there's a historical reason to keep the old snapshot, I recommend bringing the test data up to date with the current state of NCBI's taxonomy database.

This would be a great use for a CI integration, if you're feeling up for it!
