### 2024-12-20 #7604 @ryneches
A few minor issues so far :
- [Fix typo in CLI docs](https://github.com/ebi-jdispatcher/taxonomy-resolver/pull/12)
- [Bring test data up to date](https://github.com/ebi-jdispatcher/taxonomy-resolver/issues/13)
