### 2024-12-20 a4426c8
updates taxonomy dump and updates the size of trees
