+++ tests/test_tree.py
@@ -40,13 +40,13 @@ class TestTree:
         resolver = TaxonResolver(logging=context)
         resolver.build(os.path.join(cwd, "../testdata/taxdmp.zip"))
         if resolver.tree is not None:
-            assert len(resolver.tree) == 2302938
+            assert len(resolver.tree) == 2631861
 
     def test_resolver_build_and_write(self, context, cwd):
         resolver = TaxonResolver(logging=context)
         resolver.build(os.path.join(cwd, "../testdata/taxdmp.zip"))
         if resolver.tree is not None:
-            assert len(resolver.tree) == 2302938
+            assert len(resolver.tree) == 2631861
         resolver.write(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         assert os.path.isfile(os.path.join(cwd, "../testdata/tree.pickle"))
 
@@ -54,21 +54,21 @@ class TestTree:
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         if resolver.tree is not None:
-            assert len(resolver.tree) == 2302938
+            assert len(resolver.tree) == 2631861
 
     def test_resolver_filter(self, context, cwd):
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         resolver.filter(os.path.join(cwd, "../testdata/taxids_filter.txt"))
         if resolver.tree is not None:
-            assert len(resolver.tree) == 1000
+            assert len(resolver.tree) == 1003
 
     def test_resolver_filter_and_write(self, context, cwd):
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         resolver.filter(os.path.join(cwd, "../testdata/taxids_filter.txt"))
         if resolver.tree is not None:
-            assert len(resolver.tree) == 1000
+            assert len(resolver.tree) == 1003
         resolver.write(os.path.join(cwd, "../testdata/tree_filtered.pickle"), "pickle")
         assert os.path.isfile(os.path.join(cwd, "../testdata/tree_filtered.pickle"))
 
@@ -76,7 +76,7 @@ class TestTree:
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree_filtered.pickle"), "pickle")
         if resolver.tree is not None:
-            assert len(resolver.tree) == 1000
+            assert len(resolver.tree) == 1003
 
     def test_resolver_search_by_taxid_human(self, context, cwd):
         resolver = TaxonResolver(logging=context)
@@ -90,63 +90,63 @@ class TestTree:
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         taxids = resolver.search(["2"])
         if taxids:
-            assert len(taxids) == 517912
+            assert len(taxids) == 577322
 
     def test_resolver_search_by_taxid_archaea(self, context, cwd):
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         taxids = resolver.search(["2157"])
         if taxids:
-            assert len(taxids) == 13683
+            assert len(taxids) == 14866
 
     def test_resolver_search_by_taxid_eukaryota(self, context, cwd):
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         taxids = resolver.search(["2759"])
         if taxids:
-            assert len(taxids) == 1541629
+            assert len(taxids) == 1760410
 
[…]
