### 2024-12-20 ebi-jdispatcher/taxonomy-resolver#13 @biomadeira
That's a good point! There is no reason for that so I have pushed a couple of commits to update it (see [!14](https://github.com/ebi-jdispatcher/taxonomy-resolver/pull/14)). Also updating the `taxids_filter.txt` to match the new classification.

The main problem with updating the dump from CI/CD is that tree sizes are the main assertion performed throughout test cases, and these depend on the heirarchy being used. Surprisingly (or not) the NCBI Taxonomy Hierarchy does change frequently. The dump is update every few days or so.
