### 2024-12-23 ebi-jdispatcher/taxonomy-resolver#13 @ryneches
Ah, yeah. Magic numbers are maybe not the ideal way of structuring tests. Especially since the IDs are also magic numbers! It makes them very difficult to understand.

I think it would be clearer to explicitly test the general properties of trees by comparing different parts of the tree to one another. For example, the number of taxa in clade should be equal to the sum of the number of taxa in its daughter clades. Taxa that appear in one clade should not also appear in sister clades. When you pick a random set of taxa from the tree and then filter for them, they should still be in the tree afterward. If you iterate up the tree, you should always get to the root node.

They work they way they are, though, so this isn't a point I would insist on. Remembering to update magic numbers is work, though!
