### 2024-12-25 #7604 @ryneches
Looks good to me.
