### 2025-01-02 a9056db
Update README.rst to removed reference to JSON `build --outformat`
JSON outformat appears to be unsupported, so its reference on the README should go.
