### 2025-01-03 ebi-jdispatcher/taxonomy-resolver#16 @tbrittoborges
Command line interface is not covered by tests
I would suggest extending the test suite to cover the command line interface.
