### 2025-01-03 #7604 @tbrittoborges
One minor issue:
- https://github.com/ebi-jdispatcher/taxonomy-resolver/pull/15

And one suggestion:
- https://github.com/ebi-jdispatcher/taxonomy-resolver/issues/16

I would reinforce the @ryneches ' suggestions on integrating producing some NEWICK formated output or tree visualization, so the results can be rapidly checked.
