### 2025-01-03 #7604 @tbrittoborges
I just missed a way to quick check the tree / subtree - when possible. I don't think this is a strong requirement, but it could expand the user base for this package. For example, on deciding on the list of TaxIDs to filter, etc.
