### 2025-01-06 ebi-jdispatcher/taxonomy-resolver#16 @biomadeira
That's a good idea @tbrittoborges! I will work on that and open a PR soon.
