### 2025-01-06 ebi-jdispatcher/taxonomy-resolver#17 @biomadeira
Add tree visualization (possibly Newick) to aid tree checking and validation 
Look for ways to display a tree (possibly in Newick format) to help check the sub-trees.
See https://github.com/openjournals/joss-reviews/issues/7604#issuecomment-2556338253
