### 2025-01-06 ebi-jdispatcher/taxonomy-resolver#18 @biomadeira
Review progress reporting for downloads and tree traversal 
It would be nice to view some sort of progress bar whenever the user downloads the taxonomy dump using the CLI or performs some tree options, such as building, filtering and searching. As suggested by @ryneches there are options such as [tqdm](https://github.com/tqdm/tqdm), which should be easy to integrate.
See also https://github.com/openjournals/joss-reviews/issues/7604#issuecomment-2556338253
