### 2025-01-06 #7604 @biomadeira
Happy new year @ryneches @tbrittoborges!
Thanks for your suggestions and fixes!

I am picking up on this after a nice break and hoping to push PRs soon.
