### 2025-01-06 4d84b3b
test: adds missing CLI tests for most commands and options resolves #16
