### 2025-01-06 61bbcd5
test: adds missing CLI tests for most commands and options resolves #16
