+++ tests/test_cli.py
@@ -1,0 +1,301 @@
+#!/usr/bin/env python
+# -*- coding: utf-8
+
+"""
+Taxonomy Resolver
+
+:copyright: (c) 2020-2025.
+:license: Apache 2.0, see LICENSE for more details.
+"""
+
+import os
+
+import click
+import pytest
+from click.testing import CliRunner
+
+from taxonomyresolver import __version__
+from taxonomyresolver.cli import add_common, cli, common_options, common_options_parsing
+from taxonomyresolver.utils import load_logging
+
+
+@pytest.fixture
+def context():
+    return load_logging("INFO")
+
+
+@pytest.fixture
+def cwd():
+    if not os.getcwd().endswith("tests"):
+        os.chdir(os.path.join(os.getcwd(), "tests"))
+    return os.getcwd()
+
+
+@pytest.fixture
+def runner():
+    return CliRunner()
+
+
+@click.command()
+@add_common(common_options)
+@add_common(common_options_parsing)
+def demo_cli_common_options(log_level, log_output, quiet, sep, indx):
+    """A sample command to test 'add_common' decoration and 'common_options'."""
+    click.echo(
+        f"log_level={log_level}, log_output={log_output}, quiet={quiet}, "
+        f"sep={sep}, indx={indx}"
+    )
+
+
+def test_default_behavior(runner):
+    """Test the default behavior with no options passed."""
+    result = runner.invoke(demo_cli_common_options)
+    assert result.exit_code == 0
+    assert "log_level=INFO" in result.output
+    assert "log_output=None" in result.output
+    assert "quiet=False" in result.output
+    assert "sep=None" in result.output
+    assert "indx=0" in result.output
+
+
+def test_log_level_option(runner):
+    """Test setting the log_level option."""
+    result = runner.invoke(demo_cli_common_options, ["--log_level", "DEBUG"])
+    assert result.exit_code == 0
+    assert "log_level=DEBUG" in result.output
+
+
+def test_log_output_option(runner):
+    """Test setting the log_output option."""
+    result = runner.invoke(demo_cli_common_options, ["--log_output", "log.txt"])
+    assert result.exit_code == 0
+    assert "log_output=log.txt" in result.output
+
+
+def test_quiet_flag(runner):
+    """Test enabling the quiet flag."""
+    result = runner.invoke(demo_cli_common_options, ["--quiet"])
+    assert result.exit_code == 0
+    assert "quiet=True" in result.output
+
+
+def test_setting_sep(runner):
+    """Test setting sep."""
+    result = runner.invoke(demo_cli_common_options, ["--sep", "test"])
+    assert result.exit_code == 0
+    assert "sep=test" in result.output
+
+
+def test_setting_indx(runner):
+    """Test setting indx."""
+    result = runner.invoke(demo_cli_common_options, ["--indx", "1"])
+    assert result.exit_code == 0
+    assert "indx=1" in result.output
+
+
+def test_combined_options(runner):
+    """Test multiple options together."""
+    result = runner.invoke(
+        demo_cli_common_options,
+        ["--log_level", "ERROR", "--log_output", "error.log", "--quiet"],
+    )
+    assert result.exit_code == 0
+    assert "log_level=ERROR" in result.output
+    assert "log_output=error.log" in result.output
+    assert "quiet=True" in result.output
+
+
+def test_cli_help(runner):
+    """Test the CLI group help message."""
+    result = runner.invoke(cli, ["--help"])
+    assert result.exit_code == 0
+    assert "Taxonomy Resolver" in result.output
+    assert "Build a NCBI Taxonomy Tree" in result.output
+
+    result = runner.invoke(cli, ["-h"])
+    assert result.exit_code == 0
[…]
