### 2025-01-06 ebi-jdispatcher/taxonomy-resolver#13 @biomadeira
@ryneches These are good points. We do have a number of tests that work somewhat like you describe using a mocked tree, but I will work to improve the main ones.
