### 2025-01-07 0f9b946
chore: add tqdm dependency
