+++ requirements.txt
@@ -2,3 +2,4 @@ click>=8.1.7
 requests>=2.32.2
 pytest>=8.2.1
 pandas>=2.2.2
+tqdm>=4.65.0
