### 2025-01-07 0cb82d2
feat: adds progress reporting to download
