### 2025-01-07 c2944a7
feat: adds progress reporting reading the dump and building the tree structure
