### 2025-01-07 9649af6
test: improves tests so they do not rely on the size of the trees
