+++ tests/test_tree.py
@@ -40,13 +40,15 @@ class TestTree:
         resolver = TaxonResolver(logging=context)
         resolver.build(os.path.join(cwd, "../testdata/taxdmp.zip"))
         if resolver.tree is not None:
-            assert len(resolver.tree) == 2631861
+            assert len(resolver.tree) > 0
+            assert "9606" in resolver.tree["id"].values
 
     def test_resolver_build_and_write(self, context, cwd):
         resolver = TaxonResolver(logging=context)
         resolver.build(os.path.join(cwd, "../testdata/taxdmp.zip"))
         if resolver.tree is not None:
-            assert len(resolver.tree) == 2631861
+            assert len(resolver.tree) > 0
+            assert "9606" in resolver.tree["id"].values
         resolver.write(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         assert os.path.isfile(os.path.join(cwd, "../testdata/tree.pickle"))
 
@@ -54,21 +56,26 @@ class TestTree:
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         if resolver.tree is not None:
-            assert len(resolver.tree) == 2631861
+            assert len(resolver.tree) > 0
+            assert "9606" in resolver.tree["id"].values
 
     def test_resolver_filter(self, context, cwd):
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         resolver.filter(os.path.join(cwd, "../testdata/taxids_filter.txt"))
         if resolver.tree is not None:
-            assert len(resolver.tree) == 1003
+            assert len(resolver.tree) > 0
+            assert "9606" in resolver.tree["id"].values
+            assert "4751" not in resolver.tree["id"].values
 
     def test_resolver_filter_and_write(self, context, cwd):
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         resolver.filter(os.path.join(cwd, "../testdata/taxids_filter.txt"))
         if resolver.tree is not None:
-            assert len(resolver.tree) == 1003
+            assert len(resolver.tree) > 0
+            assert "9606" in resolver.tree["id"].values
+            assert "4751" not in resolver.tree["id"].values
         resolver.write(os.path.join(cwd, "../testdata/tree_filtered.pickle"), "pickle")
         assert os.path.isfile(os.path.join(cwd, "../testdata/tree_filtered.pickle"))
 
@@ -76,77 +83,89 @@ class TestTree:
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree_filtered.pickle"), "pickle")
         if resolver.tree is not None:
-            assert len(resolver.tree) == 1003
+            assert len(resolver.tree) > 0
+            assert "9606" in resolver.tree["id"].values
+            assert "4751" not in resolver.tree["id"].values
 
     def test_resolver_search_by_taxid_human(self, context, cwd):
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         taxids = resolver.search(["9606"])
         if taxids:
-            assert len(taxids) == 3
+            assert len(taxids) > 0
+            assert "9606" in taxids
 
     def test_resolver_search_by_taxid_bacteria(self, context, cwd):
         resolver = TaxonResolver(logging=context)
         resolver.load(os.path.join(cwd, "../testdata/tree.pickle"), "pickle")
         taxids = resolver.search(["2"])
         if taxids:
-            assert len(taxids) == 577322
[…]
