### 2025-01-08 efe21c7
feature: adds a new function to generate newick out of a tree structure
