### 2025-01-08 1e5229a
test: adds a new test to cover the new function and using the mock tree
