### 2025-01-08 d6f6546
feature: enables writing newick format for the build and search commands
