### 2025-01-08 a9d79b5
feature: enables writing newick format for the build and search commands
