### 2025-01-08 a35d2a7
docs: adds extra CLI examples for outputing newick format
