### 2025-01-08 ebi-jdispatcher/taxonomy-resolver#17 @biomadeira
@ryneches @tbrittoborges I have added Newick support. Have a look at the last three commands in the CLI part of the README to see what my angle was. I am happy that you suggested this, as it should be there already and it is super useful. I was quite surprised that it can build the entire NCBI Taxonomy tree in newick in a couple of minutes. It is huge, but reassuring to see it.
