### 2025-01-09 22f3f5e
chore: bumps version to 1.1.0
