### 2025-01-09 cc6ff1a
docs: noting newick as an output
