### 2025-02-01 #7604 @csoneson
Hi @biomadeira - I have gone through the submission and it looks good to me. Only two minor issues in the bibliography:

* The `BigTree` GitHub repo link is missing an initial `h`
* The `ggtree` paper title needs some case protection (for `ggtree` and `R`)

Once these are fixed, you can generate a new proof with `@editorialbot generate pdf`.

I will make another post below with the next steps.
