+++ paper/paper.bib
@@ -149,7 +149,7 @@ CONCLUSION: The new BLAST command-line applications, compared to the current BLA
 	year = {2024},
 	publisher = {GitHub},
 	journal = {GitHub repository},
-	url = {ttps://github.com/kayjan/bigtree}
+	url = {https://github.com/kayjan/bigtree}
 }
 
 @article{madeira_2024,
