+++ paper/paper.bib
@@ -38,7 +38,7 @@ CONCLUSION: The new BLAST command-line applications, compared to the current BLA
 
 @article{yu_ggtree_2017,
 	author = {Yu, Guangchuang and Smith, David K. and Zhu, Huachen and Guan, Yi and Lam, Tommy Tsan-Yuk},
-	title = {ggtree: an r package for visualization and annotation of phylogenetic trees with their covariates and other associated data},
+	title = {{ggtree}: an {R} package for visualization and annotation of phylogenetic trees with their covariates and other associated data},
 	journal = {Methods in Ecology and Evolution},
 	volume = {8},
 	number = {1},
