### 2025-02-10 #7604 @csoneson
Hi @biomadeira - thanks for fixing the references. Please see the [post](https://github.com/openjournals/joss-reviews/issues/7604#issuecomment-2628890281) above for the next steps in the process.
