+++ paper/paper.md
@@ -17,11 +17,11 @@ authors:
   - name: Nandana Madhusoodanan
     orcid: 0000-0001-5004-152X
     affiliation: 1
-  - name: Alberto Eusebi
-    orcid: 0000-0001-5179-7724
-    affiliation: 1
   - name: Joonheung Lee
     orcid: 0000-0002-5760-2761
+    affiliation: 1
+  - name: Alberto Eusebi
+    orcid: 0000-0001-5179-7724
     affiliation: 1
   - name: Ania Niewielska
     orcid: 0000-0003-0989-3389
