### 2025-02-10 #7604 @biomadeira
Hi @csoneson again. I think I have been all those.
I have double checked the authors list, their affiliations and ORCID. I created a new release [1.1.1](https://github.com/ebi-jdispatcher/taxonomy-resolver/releases/tag/1.1.1), which is published as well in [Zenodo](https://zenodo.org/records/14847194).
Let me know if there is anything else, or that I may have missed.
