### 2025-02-13 2cb8025
Use single line syntax
