### 2025-02-13 #7604 @xuanxu
The multiline syntax used to set the institution name in the metadata is not allowed, [this PR](https://github.com/ebi-jdispatcher/taxonomy-resolver/pull/23) should fix the error.
