### 2025-02-13 #7604 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
