### 2025-02-20 82f0fbb
Update paper.md
paper: fixes UK spelling openjournals/joss-reviews#7604
