### 2025-02-21 ec3273b
docs: updates paper badge
